requirements.txt
numpy
tensorflow
keras
Pillow==8.3.2
flask==2.0.1
gdown
